refactor(pages): rename folder from Home to tmp
